Додав spring-boot-maven-plugin. Тепер проект можна запускати з командного рядка 'mvn spring-boot:run'. Всі поля таблиць (models) зробив об"єктами. Id -> Long.   А також Name Convention, Code Convention, Case Sensitive Names.

+++ src/main/java/com/trello/service/BoardService.java
@@ -21,7 +21,7 @@ public class BoardService {
         repository.save(board);
     }
 
-    public Board findOne(int id ){Board board=repository.findOne(id);
+    public Board findOne(Long id){Board board=repository.findOne(id);
             return board;
     }
     public List<Board> getAll(){
